requirements.txt
discord.py
youtube_dl
python_dotenv
async_timeout
PyNaCl
requests
dataclasses_json
beautifulsoup4